requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data Validation & Configuration
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7
PyJWT==2.8.1
python-multipart==0.0.6

# Background Tasks & Scheduling
apscheduler==3.10.4

# Notifications
twilio==8.13.0

# Testing & Coverage
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
coverage==7.4.1

# Environment Configuration
python-dotenv==1.0.0
